refactor(rest): extract plugin-job and job-status/cancel routes into routes/plugin_jobs.py
Moves POST /api/plugins/{plugin_id}/jobs, POST /api/jobs/{job_id}/status,
and the per-user in-flight job list/cancel (GET/POST
/api/scopes/{scope}/my-jobs...) out of the create_app closure into a
standalone APIRouter.

enqueue_plugin_job is the one enqueue path shared with the plugin-job
scheduler's tick, which still lives in create_app (admin/plugin-jobs
schedules are not extracted yet) -- it moves to routes/plugin_jobs.py as a
plain function taking RestContext explicitly, and create_app keeps calling
it as _enqueue_plugin_job, now a thin wrapper bound to the app's
RestContext. plugin_job_requires_admin has no other caller and moves in
full. Lifts require_pool, human_bytes and pending_upload_detail onto
routes/deps.py (needed by the new module's plugin-job route; the closure
keeps its old underscore names as aliases for the routes still inside it).

Updates test_job_status_report's source-literal guard to read
routes/plugin_jobs.py, where _REPORTABLE_JOB_STATUSES now lives, instead
of app.py.
